requirements.txt
asgiref==3.7.2
attrs==23.1.0
diploma-frontend==0.6
Django==4.2.2
django-request-logging==0.7.5
djangorestframework==3.14.0
drf-spectacular==0.26.3
inflection==0.5.1
jsonschema==4.17.3
Pillow==10.0.0
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
typing_extensions==4.6.3
uritemplate==4.1.1
